@@ -1,9 +1,9 @@
 abstract class BarcodeSymbol {
   BarcodeSymbol({this.value});
 
-  final String value;
+  final String? value;
 
-  List<bool> toPattern();
+  List<bool>? toPattern();
 
   BarcodeSymbol setValue(String v);
 }
