@@ -3,11 +3,11 @@ import 'package:flutter_barcode/barcode/symbols/barcode_symbol.dart';
 
 class LeftOddParityModule extends BarcodeSymbol
     with JANBarcodeModuleSpecification {
-  LeftOddParityModule({String value}) : super(value: value);
+  LeftOddParityModule({String? value}) : super(value: value);
 
   @override
   LeftOddParityModule setValue(String v) => LeftOddParityModule(value: v);
 
   @override
-  List<bool> toPattern() => patternSpecification[value];
+  List<bool>? toPattern() => patternSpecification[value];
 }
